V.tile-ify: PointOnSurface for label placement
ST_PointOnSurface doesn't work if the geometry is invalid.
Rather than risking an SQL query error, just return NULL if the polygon
is invalid, which means no point.

There is a geometry index on way, but not on ST_PointOnSurface(way), so
by default it won't be able to use an index. By putting `way && !bbox!`
the query will use the index