V.tile-ify: Disable polygon simplification at z14 level
Without this change, the z14 tiles you generate will be fine for
rendering at z14, but when you use overzooming, the polygons will be
noticably simplified.

However if tilelive sees that you're generating a vector tile at the
maximum zoom level, it will change the simplification from 8 to 1 (which
essentially means no simplification). cf commit 2a63c77 in
tilelive-bridge[1].

Hence we need to tell tilelive that 14 is the max zoom. You need to
change the toplevel maxzoom key, and have a maxzoom key in all the other
datasources (or it defaults to 22 for each one).

Since you have to change the toplevel maxzoom to 14, and you'll want to
renderin z>14, you can't use the same project.ya?ml file for creating
pbfs and raster tiles.

[1]
https://github.com/mapbox/tilelive-bridge/commit/2a63c775557d22abb40fa554ab4ef9bd03fc5ec1