V.tile-ify: Make way_pixels work
way_pixels is used for deciding whether an object should be shown. It
adapts to the zoom level. However it's calculated in data, so for the
z14 vectortiles, the way_pixels size is fixed at the z14 size.

This means some style rules won't be applied for z14+ tiles.

Solution is to manually add in some new way_pixel filters based on the
z14 filter, and taking into account the fact that one pixel in Z is 4
pixels in Z+1.