Remove some styling (about `way_pixels`) to get carto performance sensible
Previously, this style had attempted to replicate the upstream osm-carto
exactly. This meant the using the `way_pixels` and zoom trick to get the
same results. But that results in a combinatorial explosion for
amenity-points, which means carto can go from 30 seconds to parse the
file, to 3½minutes. By removing this inner selectors, the carto
performance will be back to about 30 seconds.

This affects the start up time of tessera/tilelive.

cf.:

 * https://github.com/mapbox/carto/issues/470
 * https://github.com/mapbox/carto/issues/20